+++ source/binding/ConstantValue.cpp
@@ -16,37 +16,38 @@ namespace slang {
 const ConstantValue ConstantValue::Invalid;
 
 std::string ConstantValue::toString() const {
-    return std::visit([](auto&& arg) noexcept {
-        using T = std::decay_t<decltype(arg)>;
-        if constexpr (std::is_same_v<T, std::monostate>)
-            return "<unset>"s;
-        else if constexpr (std::is_same_v<T, SVInt>)
-            return arg.toString();
-        else if constexpr (std::is_same_v<T, real_t>)
-            return std::to_string(arg);
-        else if constexpr (std::is_same_v<T, shortreal_t>)
-            return std::to_string(arg);
-        else if constexpr (std::is_same_v<T, ConstantValue::NullPlaceholder>)
-            return "null"s;
-        else if constexpr (std::is_same_v<T, Elements>) {
-            FormatBuffer buffer;
-            buffer.append("[");
-            for (auto& element : arg) {
-                buffer.append(element.toString());
-                buffer.append(",");
-            }
+    return std::visit(
+        [](auto&& arg) noexcept {
+            using T = std::decay_t<decltype(arg)>;
+            if constexpr (std::is_same_v<T, std::monostate>)
+                return "<unset>"s;
+            else if constexpr (std::is_same_v<T, SVInt>)
+                return arg.toString();
+            else if constexpr (std::is_same_v<T, real_t>)
+                return std::to_string(arg);
+            else if constexpr (std::is_same_v<T, shortreal_t>)
+                return std::to_string(arg);
+            else if constexpr (std::is_same_v<T, ConstantValue::NullPlaceholder>)
+                return "null"s;
+            else if constexpr (std::is_same_v<T, Elements>) {
+                FormatBuffer buffer;
+                buffer.append("[");
+                for (auto& element : arg) {
+                    buffer.append(element.toString());
+                    buffer.append(",");
+                }
 
-            if (!arg.empty())
-                buffer.pop_back();
-            buffer.append("]");
-            return buffer.str();
-        }
-        else if constexpr (std::is_same_v<T, std::string>)
-            return arg;
-        else
-            static_assert(always_false<T>::value, "Missing case");
-    },
-                      value);
+                if (!arg.empty())
+                    buffer.pop_back();
+                buffer.append("]");
+                return buffer.str();
+            }
+            else if constexpr (std::is_same_v<T, std::string>)
+                return arg;
+            else
+                static_assert(always_false<T>::value, "Missing case");
+        },
+        value);
 }
 
 ConstantValue ConstantValue::getSlice(int32_t upper, int32_t lower) const {
@@ -54,51 +55,53 @@ ConstantValue ConstantValue::getSlice(int32_t upper, int32_t lower) const {
         return integer().slice(upper, lower);
 
     if (isUnpacked()) {
-        auto slice = elements().subspan(lower, upper - lower + 1);
+        auto slice = elements().subspan(size_t(lower), size_t(upper - lower + 1));
         return std::vector<ConstantValue>(slice.begin(), slice.end());
     }
 
     if (isString()) {
         ASSERT(upper == lower);
         ASSERT(upper >= 0);
-        return SVInt(8, (uint64_t)str()[uint64_t(upper)], false);
+        return SVInt(8, (uint64_t)str()[size_t(upper)], false);
     }
 
     return nullptr;
 }
 
 bool ConstantValue::isTrue() const {
-    return std::visit([](auto&& arg) noexcept {
-        using T = std::decay_t<decltype(arg)>;
-        if constexpr (std::is_same_v<T, SVInt>)
-            return (bool)(logic_t)arg;
-        else if constexpr (std::is_same_v<T, real_t>)
-            return (bool)arg;
-        else if constexpr (std::is_same_v<T, shortreal_t>)
-            return (bool)arg;
-        else
-            return false;
-    },
-                      value);
+    return std::visit(
+        [](auto&& arg) noexcept {
+            using T = std::decay_t<decltype(arg)>;
+            if constexpr (std::is_same_v<T, SVInt>)
+                return (bool)(logic_t)arg;
+            else if constexpr (std::is_same_v<T, real_t>)
+                return (bool)arg;
+            else if constexpr (std::is_same_v<T, shortreal_t>)
+                return (bool)arg;
+            else
+                return false;
+        },
+        value);
 }
 
 bool ConstantValue::isFalse() const {
-    return std::visit([](auto&& arg) noexcept {
-        using T = std::decay_t<decltype(arg)>;
-        if constexpr (std::is_same_v<T, SVInt>) {
-            logic_t l = (logic_t)arg;
-            return !l.isUnknown() && l.value == 0;
-        }
-        else if constexpr (std::is_same_v<T, real_t>)
-            return !(bool)arg;
-        else if constexpr (std::is_same_v<T, shortreal_t>)
-            return !(bool)arg;
-        else if constexpr (std::is_same_v<T, ConstantValue::NullPlaceholder>)
-            return true;
-        else
-            return false;
-    },
-                      value);
+    return std::visit(
+        [](auto&& arg) noexcept {
+            using T = std::decay_t<decltype(arg)>;
+            if constexpr (std::is_same_v<T, SVInt>) {
+                logic_t l = (logic_t)arg;
+                return !l.isUnknown() && l.value == 0;
+            }
+            else if constexpr (std::is_same_v<T, real_t>)
+                return !(bool)arg;
+            else if constexpr (std::is_same_v<T, shortreal_t>)
+                return !(bool)arg;
+            else if constexpr (std::is_same_v<T, ConstantValue::NullPlaceholder>)
+                return true;
+            else
+                return false;
+        },
+        value);
 }
 
 bool ConstantValue::equivalentTo(const ConstantValue& rhs) const {
@@ -120,11 +123,11 @@ bool ConstantValue::equivalentTo(const ConstantValue& rhs) const {
                     return false;
 
                 auto rhsElements = rhs.elements();
-                if (arg.size() != (size_t)rhsElements.size())
+                if (arg.size() != rhsElements.size())
                     return false;
 
                 for (size_t i = 0; i < arg.size(); i++) {
-                    if (!arg[i].equivalentTo(rhsElements[ptrdiff_t(i)]))
+                    if (!arg[i].equivalentTo(rhsElements[i]))
                         return false;
                 }
 
@@ -261,25 +264,26 @@ std::ostream& operator<<(std::ostream& os, const ConstantRange& cr) {
 }
 
 ConstantValue LValue::load() const {
-    return std::visit([](auto&& arg) noexcept(!std::is_same_v<std::decay_t<decltype(arg)>, Concat>)
-                          ->ConstantValue {
-                              using T = std::decay_t<decltype(arg)>;
-                              if constexpr (std::is_same_v<T, std::monostate>)
-                                  return ConstantValue();
-                              else if constexpr (std::is_same_v<T, ConstantValue*>)
-                                  return *arg;
-                              else if constexpr (std::is_same_v<T, CVRange>)
-                                  return arg.cv->getSlice(arg.range.upper(), arg.range.lower());
-                              else if constexpr (std::is_same_v<T, Concat>) {
-                                  SmallVectorSized<SVInt, 4> vals;
-                                  for (auto& elem : arg)
-                                      vals.append(elem.load().integer());
-                                  return SVInt::concat(vals);
-                              }
-                              else
-                                  static_assert(always_false<T>::value, "Missing case");
-                          },
-                      value);
+    return std::visit(
+        [](auto&& arg) noexcept(!std::is_same_v<std::decay_t<decltype(arg)>, Concat>)
+            ->ConstantValue {
+                using T = std::decay_t<decltype(arg)>;
+                if constexpr (std::is_same_v<T, std::monostate>)
+                    return ConstantValue();
+                else if constexpr (std::is_same_v<T, ConstantValue*>)
+                    return *arg;
+                else if constexpr (std::is_same_v<T, CVRange>)
+                    return arg.cv->getSlice(arg.range.upper(), arg.range.lower());
+                else if constexpr (std::is_same_v<T, Concat>) {
+                    SmallVectorSized<SVInt, 4> vals;
+                    for (auto& elem : arg)
+                        vals.append(elem.load().integer());
+                    return SVInt::concat(vals);
+                }
+                else
+                    static_assert(always_false<T>::value, "Missing case");
+            },
+        value);
 }
 
 void LValue::store(const ConstantValue& newValue) {
@@ -305,14 +309,14 @@ void LValue::store(const ConstantValue& newValue) {
                     auto dest = cv.elements();
 
                     for (int32_t i = l; i <= u; i++)
-                        dest[i] = src[i - l];
+                        dest[size_t(i)] = src[size_t(i - l)];
                 }
                 else if (cv.isString()) {
                     ASSERT(l == u);
                     ASSERT(l >= 0);
                     char c = (char)*newValue.integer().as<uint8_t>();
                     if (c)
-                        cv.str()[uint64_t(l)] = c;
+                        cv.str()[size_t(l)] = c;
                 }
                 else {
                     cv.integer().set(u, l, newValue.integer());
@@ -361,9 +365,9 @@ LValue LValue::selectIndex(int32_t index) const {
             if constexpr (std::is_same_v<T, std::monostate>)
                 return nullptr;
             else if constexpr (std::is_same_v<T, ConstantValue*>)
-                return LValue(arg->elements()[index]);
+                return LValue(arg->elements()[size_t(index)]);
             else if constexpr (std::is_same_v<T, CVRange>)
-                return LValue(arg.cv->elements()[arg.range.lower() + index]);
+                return LValue(arg.cv->elements()[size_t(arg.range.lower() + index)]);
             else if constexpr (std::is_same_v<T, Concat>)
                 THROW_UNREACHABLE;
             else
